Add auth_token logic to auth views
Prior to fulfilling each request, a check is performed to ensure that
the auth_token stored in the session cookie matches the auth_token
in the User model of the current user. If they don't match, the session
is logged off. This ensures that if a user's session cookie is
stolen, the user can change either their account password or email
address in order to invalidate the stolen cookie. Thus, the stolen
cookie will no longer allow the attacker to access the user's account.